Add `MatchesExact` function
This function allows matching on the full path, without attempting to
match any parent elements of the path.

This was *technically* possible before, by calling the deprecated
`MatchesUsingParentResult` and always setting `parentMatched` to
`false`. However, this is quite hacky and results in quite
tricky-to-read code (and also didn't have tests, etc). So this patch
adds in a proper function for this, and ensures it works with some
refactored tests.

Signed-off-by: Justin Chadwell <me@jedevc.com>